refactor: name _socket to socket
