+++ CMakeLists.txt
@@ -27,8 +27,8 @@ add_library(${PROJECT_NAME} SHARED
   "src/${PROJECT_NAME}/communication/tcp/receiver.cpp"
   "src/${PROJECT_NAME}/communication/tcp/sender.cpp"
   "src/${PROJECT_NAME}/communication/tcp/client.cpp"
-  "src/${PROJECT_NAME}/communication/_socket/base_socket.cpp"
-  "src/${PROJECT_NAME}/communication/_socket/tcp_socket.cpp"
+  "src/${PROJECT_NAME}/communication/socket/base_socket.cpp"
+  "src/${PROJECT_NAME}/communication/socket/tcp_socket.cpp"
   "src/${PROJECT_NAME}/node/robocup_client_node.cpp"
   "src/${PROJECT_NAME}/robot_client/receiver.cpp"
   "src/${PROJECT_NAME}/robot_client/sender.cpp"
