--- include/robocup_client/communication/_socket/base_socket.hpp
@@ -1,51 +1,0 @@
-// Copyright (c) 2021 ICHIRO ITS
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-// THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-#ifndef ROBOCUP_CLIENT__COMMUNICATION__SOCKET__BASE_SOCKET_HPP_
-#define ROBOCUP_CLIENT__COMMUNICATION__SOCKET__BASE_SOCKET_HPP_
-
-namespace robocup_client
-{
-
-namespace communication
-{
-
-class BaseSocket
-{
-public:
-  BaseSocket();
-  ~BaseSocket();
-
-  virtual bool connect();
-  virtual bool disconnect();
-
-  const int & get_sockfd() const;
-
-  bool is_connected() const;
-
-protected:
-  int sockfd;
-};
-
-}  // namespace communication
-
-}  // namespace robocup_client
-
-#endif  // ROBOCUP_CLIENT__COMMUNICATION__SOCKET__BASE_SOCKET_HPP_
